Move `TargetPath` into its own package
