+++ pkg/driver/node/targetpath/target_path.go
@@ -1,4 +1,4 @@
-package node
+package targetpath
 
 import (
 	"errors"
@@ -6,7 +6,7 @@ import (
 )
 
 var (
-	ErrInvalidTargetPath = errors.New("ParseTargetPath: Invalid target path")
+	ErrInvalidTargetPath = errors.New("targetpath: parse: Invalid target path")
 )
 
 var (
@@ -21,10 +21,10 @@ type TargetPath struct {
 	VolumeID string
 }
 
-// ParseTargetPath parses given target path from Kubernetes.
+// Parse parses given target path from Kubernetes.
 // Target paths are generated by Kubernetes for CSI drivers to mount their filesystem.
 // They are in the form of "/var/lib/kubelet/pods/<pod-uuid>/volumes/kubernetes.io~csi/<volume-id>/mount".
-func ParseTargetPath(path string) (TargetPath, error) {
+func Parse(path string) (TargetPath, error) {
 	matches := targetPathRegexp.FindStringSubmatch(path)
 	if len(matches) < targetPathRegexpVolumeIdIndex {
 		return TargetPath{}, ErrInvalidTargetPath
